[Clang][Driver] Revise Cygwin ToolChain to call linker directly
...so that libc++, compiler-rt, and libunwind can be used by the
options: -stdlib=libc++, -rtlib=compiler-rt, and -unwindlib=libunwind
respectively. Along with this change, the test for this driver is also
trimmed a bit.

This is a follow-up patch to the commit 52924a2d7255.

Signed-off-by: Takashi Yano <takashi.yano@nifty.ne.jp>

Misc fixes while testting 32-bit

Don't add --disable-high-entropy-va on 32-bit, it's only supported on 64-bit

For executables, add --large-address-aware for 32-bit, and --tsaware for all, to match gcc.

Map -rdynamic to --export-all-symbols to match https://cygwin.com/cgit/cygwin-packages/gcc/tree/0004-Cygwin-add-dummy-pthread-tsaware-and-large-address-a.patch

Co-authored-by: Jeremy Drake <github@jdrake.com>

add /lib/win32api

Co-authored-by: kikairoya <kikairoya@gmail.com>

adopt HIPRuntimes changes

Co-authored-by: kikairoya <kikairoya@gmail.com>